% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{T test function}
\usage{
t_test(x, y)
}
\arguments{
\item{x}{Variable x}

\item{y}{Variable y}
}
\description{
Runs an independent samples t-test using our Megabank dataset.
}
\details{
Takes the variables and returns some plots and results
}
